\name{SampleData}
\docType{data}
\alias{SampleData}
\title{Sample time series of commercial fish landings of selected marine resources (2007-2021)}
\description{A dataframe named "SampleData" was taken for illustration. It contains time series of landings of 25 resources.
}
\usage{SampleData}

\format{This data frame contains 15 rows and 26 variables}

\source{National Marine Fishery Resources Data Centre (NMFDC) of CMFRI, Kochi}

\references{
http://www.cmfri.org.in/fish-catch-estimates
}

\keyword{datasets}
